requirements.txt
fastapi 
uvicorn
pypdf2
sentence-transformers
faiss-cpu
groq
google-generativeai
python-dotenv